LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.arc.selfintersect.2.worker-expected.txt


PASS arc() with lineWidth > 2*radius is drawn sensibly 
